Refs #15014: Moved the test from StatisticsDomainParticipanTests to StatisticsQosTests.
